Refactor - revamp use of prompt revisions in playground
