+++ agenta-web/src/components/Playground/Views/TestView.tsx
@@ -1,7 +1,7 @@
 import React, {useContext, useEffect, useRef, useState} from "react"
 import {Button, Input, Card, Row, Col, Space, Form, Modal} from "antd"
 import {CaretRightOutlined, CloseCircleOutlined, PlusOutlined} from "@ant-design/icons"
-import {callVariant} from "@/lib/services/api"
+import {callVariant, promptRevision} from "@/lib/services/api"
 import {
     ChatMessage,
     ChatRole,
@@ -31,9 +31,6 @@ import duration from "dayjs/plugin/duration"
 import {useQueryParam} from "@/hooks/useQuery"
 import {dynamicComponent} from "@/lib/helpers/dynamic"
 
-const PromptVersioningDrawer: any = dynamicComponent(
-    `PromptVersioningDrawer/PromptVersioningDrawer`,
-)
 
 dayjs.extend(relativeTime)
 dayjs.extend(duration)
@@ -143,15 +140,8 @@ interface TestViewProps {
     isChatVariant?: boolean
     compareMode: boolean
     onStateChange: (isDirty: boolean) => void
-    promptRevisions: IPromptVersioning | undefined
-    historyStatus: {
-        loading: boolean
-        error: boolean
-    }
     setPromptOptParams: React.Dispatch<React.SetStateAction<Parameter[] | null>>
     promptOptParams: Parameter[] | null
-    setIsDrawerOpen: React.Dispatch<React.SetStateAction<boolean>>
-    isDrawerOpen: boolean
 }
 
 interface BoxComponentProps {
@@ -330,10 +320,6 @@ const App: React.FC<TestViewProps> = ({
     compareMode,
     onStateChange,
     setPromptOptParams,
-    promptRevisions,
-    historyStatus,
-    isDrawerOpen,
-    setIsDrawerOpen,
 }) => {
     const router = useRouter()
     const appId = router.query.app_id as unknown as string
@@ -348,7 +334,6 @@ const App: React.FC<TestViewProps> = ({
     const [resultsList, setResultsList] = useState<string[]>(testList.map(() => ""))
     const [params, setParams] = useState<Record<string, string> | null>(null)
     const classes = useStylesApp({themeMode: appTheme} as StyleProps)
-    const filteredRevisions = promptRevisions?.revisions.filter((item) => item.revision !== 0)
 
     const rootRef = React.useRef<HTMLDivElement>(null)
     const [isLLMProviderMissingModalOpen, setIsLLMProviderMissingModalOpen] = useState(false)
@@ -364,41 +349,45 @@ const App: React.FC<TestViewProps> = ({
 
     useEffect(() => {
         if (!revisionNum) return
-
-        const revision = filteredRevisions?.find((rev) => rev.revision === parseInt(revisionNum))
-
-        if (!revision) return
-
-        setPromptOptParams((prevState: Parameter[] | null) => {
-            if (!prevState) {
-                return prevState
-            }
-
-            const parameterNames = [
-                "temperature",
-                "model",
-                "max_tokens",
-                "prompt_system",
-                "prompt_user",
-                "top_p",
-                "frequence_penalty",
-                "presence_penalty",
-                "inputs",
-            ]
-
-            return prevState.map((param: Parameter) => {
-                if (parameterNames.includes(param.name)) {
-                    const newValue = (revision?.config.parameters as Record<string, any>)[
-                        param.name
-                    ]
-                    if (newValue !== undefined) {
-                        param.default = newValue
-                    }
+    
+        const fetchData = async () => {
+            const revision = await promptRevision(variant.variantId, parseInt(revisionNum))
+            if (!revision) return
+    
+            setPromptOptParams((prevState: Parameter[] | null) => {
+                if (!prevState) {
+                    return prevState
                 }
-                return param
+    
+                const parameterNames = [
+                    "temperature",
+                    "model",
+                    "max_tokens",
+                    "prompt_system",
+                    "prompt_user",
+                    "top_p",
+                    "frequence_penalty",
+                    "presence_penalty",
+                    "inputs",
+                ]
+    
+                return prevState.map((param: Parameter) => {
+                    if (parameterNames.includes(param.name)) {
+                        const newValue = (revision?.config.parameters as Record<string, any>)[
+                            param.name
+                        ]
+                        if (newValue !== undefined) {
+                            param.default = newValue
+                        }
+                    }
+                    return param
+                })
             })
-        })
+        }
+    
+        fetchData()
     }, [revisionNum])
+    
 
     const abortControllersRef = useRef<AbortController[]>([])
     const [isRunningAll, setIsRunningAll] = useState(false)
@@ -684,14 +673,6 @@ const App: React.FC<TestViewProps> = ({
                     have a valid API key for the model you are using.
                 </p>
             </Modal>
-            <PromptVersioningDrawer
-                setIsDrawerOpen={setIsDrawerOpen}
-                setRevisionNum={setRevisionNum}
-                isDrawerOpen={isDrawerOpen}
-                historyStatus={historyStatus}
-                promptRevisions={filteredRevisions}
-                onStateChange={onStateChange}
-            />
         </div>
     )
 }
